requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.9
chromadb
ollama
fastapi
uvicorn
sentence-transformers
pydantic
mcp
sqlalchemy
pymysql
python-dotenv